f32/f16 -> fp32/fp16

+++ models/turbine_models/tests/sd_test.py
@@ -28,7 +28,7 @@ arguments = {
     "batch_size": 1,
     "height": 512,
     "width": 512,
-    "precision": "fp16",
+    "precision": "fp32",
     "max_length": 77,
     "guidance_scale": 7.5,
     "run_vmfb": True,
